+++ scripts/current.py
